#51 fix eslint warnings in ThemeProvider
