+++ frontend/app/src/components/hooks/useTheme.ts
@@ -1,0 +1,24 @@
+import { Theme } from "@/enums/Theme";
+import { createContext, useContext } from "react";
+
+type ThemeProviderState = {
+	theme: Theme;
+	setTheme: (theme: Theme) => void;
+};
+
+const initialState: ThemeProviderState = {
+	theme: "system",
+	setTheme: () => null,
+};
+
+export const ThemeProviderContext =
+	createContext<ThemeProviderState>(initialState);
+
+export const useTheme = () => {
+	const context = useContext(ThemeProviderContext);
+
+	if (context === undefined)
+		throw new Error("useTheme must be used within a ThemeProvider");
+
+	return context;
+};
