+++ frontend/app/src/components/ui/ThemeProvider.tsx
@@ -1,24 +1,12 @@
-import { createContext, useContext, useEffect, useState } from "react";
-
-type Theme = "dark" | "light" | "system";
+import { Theme } from "@/enums/Theme";
+import { useEffect, useState } from "react";
+import { ThemeProviderContext } from "../hooks/useTheme";
 
 type ThemeProviderProps = {
 	children: React.ReactNode;
 	defaultTheme?: Theme;
 	storageKey?: string;
 };
-
-type ThemeProviderState = {
-	theme: Theme;
-	setTheme: (theme: Theme) => void;
-};
-
-const initialState: ThemeProviderState = {
-	theme: "system",
-	setTheme: () => null,
-};
-
-const ThemeProviderContext = createContext<ThemeProviderState>(initialState);
 
 export function ThemeProvider({
 	children,
@@ -62,13 +50,3 @@ export function ThemeProvider({
 		</ThemeProviderContext.Provider>
 	);
 }
-//this is an shadcn components and works as inteded, their warnings can be ignored
-// eslint-disable-next-line react-refresh/only-export-components
-export const useTheme = () => {
-	const context = useContext(ThemeProviderContext);
-
-	if (context === undefined)
-		throw new Error("useTheme must be used within a ThemeProvider");
-
-	return context;
-};
