+++ frontend/app/src/enums/Theme.ts
@@ -1,0 +1,1 @@
+export type Theme = "dark" | "light" | "system";
